requirements.txt
behave-django==1.1.0
Django==1.10.3
djangorestframework==3.9.1
enum34==1.1.6
funcsigs==1.0.2
linecache2==1.0.0
mock==2.0.0
parse==1.9.0
parse-type==0.4.2
pbr==5.1.1
pprint==0.1
pytz==2018.9
traceback2==1.4.0
cucumber-tag-expressions >= 1.1.1
six >= 1.12.0
win_unicode_console; python_version < '3.6'
colorama